% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_pcaprojection.R
\name{c_pcaproj}
\alias{c_pcaproj}
\title{PCA projection}
\usage{
c_pcaproj(lp)
}
\arguments{
\item{lp}{a list of parameters created using the `validate_json_file` function

"filename": <string, required data file including more than 3 columns>

"variables": <array, numeric columns for applying PCA>

"colour": <string, categorical variable to colour the projected points>

"scale": <boolean, required whether to scale the selected variables>

"biplot": <boolean, Display biplot (loadings)>

"title": <string, title of the plot>

"caption": <string, caption of the plot>

"save": <object, composed of 'save', 'height', 'width' and 'device'>

"save": <boolean, save the file?

"height": <number, in cm of the output visualization file>

"width": <number, in cm of the output visualization file>

"device": <enum, ["eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg"]>
"interactive": <boolean, save interactive html version>

To validate the JSON object use "pca_projection_schema.json" when calling the `validate_parameters` function.}
}
\value{
a ggplot object
}
\description{
Projection based on Principal Component Analysis
}
