% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_violin.R
\name{c_violin}
\alias{c_violin}
\title{Violin plots}
\usage{
c_violin(lparams)
}
\arguments{
\item{lparams}{a list of parameters created using the `validate_json_file` function

"filename": <string, required>

"variables": <array, column names to be loaded>

"y_variable": <string, required variable to be plotted on the vertical direction>

"x_variable: <string, preferable a categorical variable>

"factorx": <boolean, whether to convert x_variable into a categorical variable>

"position" : <enum, ["dodge", "identity", "dodge2"]>

"group": <string, column variable>

"colour": <string, a column variable or a predefined color in colors()>

"fill": <string, column variable>

"linetype": <enum, ["blank", "solid", "dashed", "dotted", "dotdash", "longdash", "twodash"]>

"size": <number, line size>

"weight": <number, but no used yet>

"facet_row": <string, variable name>

"facet_column": <string, variable name>

"alpha": <number, between 0 and 1>

"color_manual": <object, composed of 'values', 'breaks' and 'labels'. See ggplot scale_color_manual documentation.

    "values": <array, color values>

    "breaks": <array, string of breaks>

    "labels": <array, labels(must be same length as breaks)>

"boxplot": <object, composed of 'addboxplot' and 'width'>

   "addboxplot": <boolean, add box plot>

   "width": <number, width of the boxplot>

"title": <string, title of thehttps://github.com/badges/shields/issues/7583 plot>

"caption": <string, caption of the plot>

"rotxlabs": <number, rotate x labels in grades>

"save": <object, composed of 'save', 'height', 'width' and 'device'>

"save": <boolean, save the file?

"height": <number, in cm of the output visualization file>

"width": <number, in cm of the output visualization file>

"device": <enum, ["eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg"]>

"interactive": <boolean, save Interactive version>

Further information can be found in \href{https://ggplot2.tidyverse.org/reference/geom_violin.html}{geom_violin} documentation.}
}
\value{
a ggplot object and if indicated in 'lparams' stores the plot in a file (s)
}
\description{
Creates violin plot(s) using ggplot
}
