% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{create_json}
\alias{create_json}
\title{Create JSON file parameters}
\usage{
create_json(
  jsonschema,
  jsonfile = NULL,
  overwrite = FALSE,
  package = "rvispack"
)
}
\arguments{
\item{jsonschema}{a JSON schema filename. Run the 'vschemas()' function to get a list of
available JSON schemas.}

\item{jsonfile}{string JSON filename to store the JSON structure and default parameters.}

\item{overwrite}{boolean flag to overwrite the 'jsonfile'.}

\item{package}{R package to which the 'jsonschema' belongs.}
}
\value{
boolean
}
\description{
Create JSON file parameters
}
