% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rugplot.R
\name{create_rug_plot}
\alias{create_rug_plot}
\title{Create a \code{rug} visualization plot}
\usage{
create_rug_plot(lp, visplot, verbose = TRUE)
}
\arguments{
\item{lp}{List of \code{rug} parameters containing information to create the plot}

\item{visplot}{A \code{rug} plot}

\item{verbose}{Additional information}
}
\value{
A ggplot object
}
\description{
Create a \code{rug} visualization plot
}
