% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{create_rugjson}
\alias{create_rugjson}
\title{Create a \code{rugplot} JSON template file}
\usage{
create_rugjson(
  visplot,
  jsonfile = NULL,
  overwrite = FALSE,
  package = "rugplot"
)
}
\arguments{
\item{visplot}{Character, a kind of visualization supported by \code{rug plot}. Run the
\code{list_rugplots()} function to get a list of available visualizations.}

\item{jsonfile}{Character, a JSON filename to store the JSON structure and default parameters.}

\item{overwrite}{Boolean, a flag to overwrite the 'JSON file'.}

\item{package}{R package that contains the JSON schema to validate the JSON file.}
}
\value{
Character, name of the created JSON file
}
\description{
This file contains the default and required parameters to create a visualization plot
using the \code{rugplot} package. The required parameters must be provided by the user.
}
