% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rugplot.R
\name{create_rugplot}
\alias{create_rugplot}
\title{Create a \code{rug} visualization plot}
\usage{
create_rugplot(lp, visplot, verbose = TRUE)
}
\arguments{
\item{lp}{List of \code{rug} parameters containing information to create the plot}

\item{visplot}{character, run \code{list_rugplots()} to see the available \code{rug} plots}

\item{verbose}{Additional information}
}
\value{
A ggplot object
}
\description{
Create visualization plot using a list of parameters created by the function \code{read_rugjson()}
}
