% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_rughelp}
\alias{display_rughelp}
\title{Display help to fill in the \code{rug} JSON template}
\usage{
display_rughelp(visplot)
}
\arguments{
\item{visplot}{character, run \code{list_rugplots()} to see the available \code{rug} plots}
}
\description{
Display detailed information about the \code{rug} parameters
}
