% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_schema}
\alias{display_schema}
\title{Display json schema}
\usage{
display_schema(jsonschema)
}
\arguments{
\item{jsonschema}{a JSON file stored in ext/data}
}
\value{
displays detailed information about the parameters required in a JSON object.
}
\description{
Display json schema
}
