% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{jschema}
\alias{jschema}
\title{Valid \code{rug} plot}
\usage{
jschema(visplot)
}
\arguments{
\item{visplot}{A \code{rug} plot.

Verify that \code{visplot} is a valid \code{rug} plot. Run \code{list_rugplots()} to see valid \code{rug} plots.}
}
\value{
Character, a JSON schema file name
}
\description{
Valid \code{rug} plot
}
\keyword{internal}
