% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{list_rugplots}
\alias{list_rugplots}
\title{List the available \code{rug} plots}
\usage{
list_rugplots()
}
\value{
Character vector including the available \code{rug} plots
}
\description{
List the available \code{rug} plots
}
