% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rug_json}
\alias{read_rug_json}
\title{Reads the \code{rug} plot parameters from a JSON file}
\usage{
read_rug_json(jsonparams, visplot, cds_package = "rugplot")
}
\arguments{
\item{jsonparams}{A JSON file to be validated against a schema}

\item{visplot}{character, a \code{rug} plot}

\item{cds_package}{The package that contains the JSON schema}
}
\description{
Reads the \code{rug} plot parameters from a JSON file
}
