% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rugjson}
\alias{read_rugjson}
\title{Read a JSON file}
\usage{
read_rugjson(jsonparams, visplot, cds_package = "rugplot")
}
\arguments{
\item{jsonparams}{Character, a \code{rug} JSON file which can be created using the
\code{create_rugjson(...)} function}

\item{visplot}{Character, a kind of visualization plot. Run \code{list_rugplots()}
to see the available plots in \code{rugplot}}

\item{cds_package}{The package that contains a JSON schema used to validate the JSON file}
}
\value{
A list of parameters to create \code{rugplot} visualization
}
\description{
This file has parameters to create a  \code{rug} visualization plot.
}
