% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_boxplot.R
\name{rug_boxplot}
\alias{rug_boxplot}
\title{Boxplot}
\usage{
rug_boxplot(lp)
}
\arguments{
\item{lp}{a list of parameters derived from a JSON object}
}
\value{
object
}
\description{
\code{rug_boxplot} is a reusable function that saves histograms in different file formats (image, pdf, interactive html).
}
\keyword{internal}
