% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_heatmap.R
\name{rug_heatmap}
\alias{rug_heatmap}
\title{Heatmap}
\usage{
rug_heatmap(lp, verbose = TRUE)
}
\arguments{
\item{lp}{List of parameters including information to create a scatter plot. The list
is created using the function \code{read_rugjson()}}

\item{verbose}{Additional information}
}
\value{
A ggplot object
}
\description{
Heatmap
}
\keyword{internal}
