% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_histogram.R
\name{rug_histogram}
\alias{rug_histogram}
\title{Histogram}
\usage{
rug_histogram(lp, verbose = TRUE)
}
\arguments{
\item{lp}{a list of parameters created from a JSON file.}

\item{verbose}{Boolean to print extended information.}
}
\value{
A ggplot object and could save the histogram in a file.
}
\description{
\code{rug_histogram} is a reusable function that saves histograms in different file formats (image, pdf, interactive html).
}
\details{
One of the best ways to create the list of parameters \code{lp} is as follows:
\enumerate{
\item Create a JSON parameters file using the \code{create_json} function.
\item Open the JSON file and fill in the required parameters.
\itemize{
\item Required parameters are in angle brackets.
\item For details of the parameters run \code{rutils::display_schema("histogram_schema.json")}.
\item Customize the rest of the pairs as needed.
\item A file name is automatically generated, If no output file name is provided.
}
\item Create and validate the list \code{lp} using  \code{validate_json_file} function.
\item Just in case, validate your JSON parameters file against the histogram schema using the \code{validate_parameters} function.
\item Run the \code{rug_histogram} function.
}
}
\examples{
# Assuming that your data file is 'iris.csv'
\dontrun{
# Step 1
create_json("histogram_schema.json")
# By default "histogram_params.json" will be generated in the current directory.

# Step 2
# Open "histogram_params.json" in your favorite editor, replace and save the following
# "<filename path>" by "iris.csv"
# "<required column name>" by "sepal_length", make sure the column name is correct
# and change '"save": false' to '"save": true'

# Step 3
lp <- rutils::validate_json_file("histogram_params.json")

# Step 4
rutils::validate_parameters("histogram_params.json","histogram_schema.json")

# Step 5
rug_histogram(lp)
# As a result, a similar 'iris.csv-hist-220118_121703.213.pdf' file will be created
}

}
