% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{rug_jsonschema}
\alias{rug_jsonschema}
\title{JSON schema for a visualization plot}
\usage{
rug_jsonschema(visplot)
}
\arguments{
\item{visplot}{}
}
\value{
string jsonschema
}
\description{
JSON schema for a visualization plot
}
