% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_parallel_coordinates.R
\name{rug_parallel}
\alias{rug_parallel}
\title{Parallel coordinates plot}
\usage{
rug_parallel(lp)
}
\arguments{
\item{lp}{A list of parameters derived from a JSON object}
}
\value{
object
}
\description{
\code{rug_parallel} is a reusable function that saves histograms in different file formats (image, pdf, interactive html).
}
\examples{
\dontrun{
create_json("parallel_schema.json")
}

}
\keyword{internal}
