% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_pcaprojection.R
\name{rug_pcaproj}
\alias{rug_pcaproj}
\title{PCA projection}
\usage{
rug_pcaproj(lp)
}
\arguments{
\item{lp}{a list of parameters created from a JSON file}
}
\value{
A ggplot object and if requested it will save the projection plot in a file.
}
\description{
\code{rug_pcaproj} is a function that creates a projection based on Principal Component Analysis
}
\details{
One of the best ways to run \code{rug_pcaproj} is as follows:
\enumerate{
\item Create a JSON parameters file using the \code{create_json} function.
\item Open the JSON file and fill in the required parameters.
      \itemize{
          \item Required parameters are in angle brackets.
          \item For details of the parameters run \code{rutils::display_schema("pca_projection_schema.json")}.
          \item Customize the rest of the pairs as needed.
          \item A file name is automatically generated, If no output file name is provided.
      }
\item Create and validate the list \code{lp} using  \code{validate_json_file} function.
\item Just in case, validate your JSON parameters file against the violin schema using the \code{validate_parameters} function.
\item Run the \code{rug_pca_proj} function.
}
}
\examples{

# Assuming that your data file is 'iris.csv'
\dontrun{
# Step 1
create_json("pca_projection_schema.json")
# By default "pca_projection_params.json" will be generated in the current directory.

# Step 2
# Open "pca_projection_params.json" in your favorite editor, replace and save the following
# "<filename path>" by "iris.csv" and the name/pair
# '"colour": null' by '"colour": "species"', make sure the column name 'species' is correct.

# Step 3
lp <- rutils::validate_json_file("pca_projection_params.json")

# Step 4
rutils::validate_parameters("pca_projection_params.json","pca_projection_schema.json")

# Step 5
rug_pcaproj(lp)
# As a result, a similar 'iris.csv-pca-220118_121703.213.pdf' file will be created.
}

}
