% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_scatter.R
\name{rug_scatter}
\alias{rug_scatter}
\title{Scatter plot}
\usage{
rug_scatter(lp, verbose = TRUE)
}
\arguments{
\item{lp}{List of parameters including information to create a scatter plot. The list
is created using the function \code{read_rugjson()}}

\item{verbose}{Additional information}
}
\value{
A ggplot object
}
\description{
Scatter plot
}
\keyword{internal}
