% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_scatterm.R
\name{rug_scatterm}
\alias{rug_scatterm}
\title{Scatter plot matrix}
\usage{
rug_scatterm(lp)
}
\arguments{
\item{lp}{}
}
\value{
object
}
\description{
\code{rug_scatterm} is a reusable function that saves scatter plot matrices in different file formats
(image, pdf, interactive html).
}
