% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_json_file}
\alias{validate_json_file}
\title{This function validates a json structure}
\usage{
validate_json_file(jsonparams)
}
\arguments{
\item{jsonparams}{Character, a JSON object file name to be validated}
}
\value{
An R list of parameters extracted from the JSON object
}
\description{
This function validates a json structure
}
\keyword{internal}
