% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate that a  json file meets a required schema}
\usage{
validate_parameters(params, pschema = "pca_projection_schema.json")
}
\arguments{
\item{params}{a json file to be validated against a schema}

\item{pschema}{a predefined json schema}
}
\description{
Validate that a  json file meets a required schema
}
\keyword{internal}
