% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putils.R
\name{create_json}
\alias{create_json}
\title{Create a \code{rug} JSON template file}
\usage{
create_json(visplot, jsonfile = NULL, overwrite = FALSE, package = "rugplot")
}
\arguments{
\item{visplot}{character, a \code{rug} plot. Run the \code{vschemas()} function to get a list of
available JSON schemas.}

\item{jsonfile}{character, a JSON filename to store the JSON structure and default parameters.}

\item{overwrite}{boolean, a flag to overwrite the 'JSON file'.}

\item{package}{R package to which \code{visplot} belongs.}
}
\value{
character, name of the created JSON file
}
\description{
Create a \code{rug} JSON template file
}
