% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rugjson}
\alias{read_rugjson}
\title{Read the \code{rug} plot parameters from a JSON file}
\usage{
read_rugjson(jsonparams, visplot, cds_package = "rugplot")
}
\arguments{
\item{jsonparams}{A \code{rug} JSON file to be validated by a JSON schema}

\item{visplot}{character, run \code{list_rugplots()} to see the available \code{rug} plots}

\item{cds_package}{The package that contains the JSON schema}
}
\value{
A list of \code{rug} parameters
}
\description{
Read the \code{rug} plot parameters from a JSON file
}
