% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_heatmap.R
\name{rug_heatmap}
\alias{rug_heatmap}
\title{Heatmap}
\usage{
rug_heatmap(lp)
}
\arguments{
\item{lp}{a list of parameters derived from a JSON object}
}
\value{
object
}
\description{
\code{rug_heatmap} is a reusable function that saves histograms in different file formats (image, pdf, interactive html).
}
