% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_scatter.R
\name{rug_scatter}
\alias{rug_scatter}
\title{Scatter plot}
\usage{
rug_scatter(lp)
}
\arguments{
\item{lp}{a list of parameters derived from a JSON object}
}
\value{
object
}
\description{
\code{rug_scatter} is a reusable function that saves scatter plots in different file formats (image, pdf, interactive html).
}
