% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_json_file}
\alias{validate_json_file}
\title{This function validates a json structure}
\usage{
validate_json_file(fileparams)
}
\arguments{
\item{fileparams}{a json structure stored in a file name to be validated}
}
\value{
an R list of parameters extracted from the json structure
}
\description{
This function validates a json structure
}
